# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Mairie des Jeunes de Covè CMS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPage
      tags: [pages]
      summary: Get page content by slug
      responses:
        "200":
          description: Page content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update page content (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageContentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageContent"

  # ===== TEAM =====
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin)
      responses:
        "204":
          description: Deleted

  # ===== NEWS =====
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      responses:
        "200":
          description: News
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getNews
      tags: [news]
      summary: Get news article
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news (admin)
      responses:
        "204":
          description: Deleted

  # ===== EVENTS =====
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List partners
      responses:
        "200":
          description: Partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create partner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update partner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete partner (admin)
      responses:
        "204":
          description: Deleted

  # ===== OPPORTUNITIES =====
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List youth opportunities
      responses:
        "200":
          description: Opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete opportunity (admin)
      responses:
        "204":
          description: Deleted

  # ===== CONTACT =====
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PageContent:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        heading: { type: string }
        subheading: { type: string }
        body: { type: string }
        heroImageUrl: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
      required: [id, slug, title]
    PageContentInput:
      type: object
      properties:
        title: { type: string }
        heading: { type: string }
        subheading: { type: string }
        body: { type: string }
        heroImageUrl: { type: string }
        seoTitle: { type: string }
        seoDescription: { type: string }
...
    Project:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        category: { type: string }
        description: { type: string }
        body: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [planned, ongoing, completed] }
        startDate: { type: string }
        endDate: { type: string }
        location: { type: string }
        budget: { type: string }
        progress: { type: integer }
        featured: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, status, createdAt]
    ProjectInput:
      type: object
      properties:
        title: { type: string }
        category: { type: string }
        description: { type: string }
        body: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [planned, ongoing, completed] }
        startDate: { type: string }
        endDate: { type: string }
        location: { type: string }
        budget: { type: string }
        progress: { type: integer }
        featured: { type: boolean }
      required: [title, status]
    NewsArticle:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        body: { type: string }
        coverImageUrl: { type: string }
        author: { type: string }
        category: { type: string }
        published: { type: boolean }
        publishedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, slug, title, published, createdAt]
    NewsInput:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        body: { type: string }
        coverImageUrl: { type: string }
        author: { type: string }
        category: { type: string }
        published: { type: boolean }
        publishedAt: { type: string, format: date-time }
      required: [title]
...
    Partner:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        logoUrl: { type: string }
        websiteUrl: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
      required: [id, name, sortOrder]
    PartnerInput:
      type: object
      properties:
        name: { type: string }
        logoUrl: { type: string }
        websiteUrl: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
      required: [name]
    Opportunity:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        category: { type: string }
        description: { type: string }
        body: { type: string }
        imageUrl: { type: string }
        deadline: { type: string }
        applyUrl: { type: string }
        location: { type: string }
        organization: { type: string }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, published, createdAt]
    OpportunityInput:
      type: object
      properties:
        title: { type: string }
        category: { type: string }
        description: { type: string }
        body: { type: string }
        imageUrl: { type: string }
        deadline: { type: string }
        applyUrl: { type: string }
        location: { type: string }
        organization: { type: string }
        published: { type: boolean }
      required: [title]
...